# lib/api-zod/src/generated/types/sellerActivityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aqary Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SellerActivityEvent {
  id: number;
  propertyId: number;
  propertyTitle: string;
  interactionType: string;
  createdAt: Date;
}