# lib/api-zod/src/generated/types/sellerActivityResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aqary Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SellerActivityEvent } from "./sellerActivityEvent";

export interface SellerActivityResponse {
  recentInteractions: SellerActivityEvent[];
}