# lib/api-zod/src/generated/types/updatePropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Aqary Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePropertyBody {
  title?: string;
  description?: string;
  price?: number;
  location?: string;
  area?: number;
  rooms?: number;
  propertyType?: string;
  features?: string[];
  imageUrl?: string;
  /** @maxItems 3 */
  imageUrls?: string[];
}